requirements.txt
Flask
psycopg2-binary
SQLAlchemy
requests
oauthlib
pyOpenSSL
pyjwt[crypto]